Address code review feedback
* split dynamodb mod into separate file
* return option from ValueBuffers consume and remove non_empty
* remove unnecessary partitions.clone in rebalance handler
